+++ src/Common/Common.cpp
@@ -1,5 +1,5 @@
 #include "Common.h"
 
-int64_t common::dist(int32_t x1, int32_t y1, int32_t x2, int32_t y2){
+double common::dist(double x1, double y1, double x2, double y2){
     return (x1 - x2) * (x1 - x2) + (y1 - y2) * (y1 - y2);
 }
